Create header file for `LanguageRegistration.cpp`
