+++ offload/languages/kernel/src/LanguageRegistration.cpp
@@ -1,4 +1,4 @@
-//===---- LanguageRegistration.cpp - Language (CUDA/HIP) registration api -===//
+//===---- LanguageRegistration.h - Language (CUDA/HIP) registration api ---===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -8,31 +8,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "ExportedAPI.h"
-
-#include "OffloadAPI.h"
-
-#include <cstdint>
-#include <iterator>
-
-#define HIP_FATBIN_MAGIC_STR "__CLANG_OFFLOAD_BUNDLE__"
-constexpr auto HIP_FATBIN_MAGIC_STR_LEN = sizeof(HIP_FATBIN_MAGIC_STR) - 1;
-
-namespace {
-struct FatbinWrapperTy {
-  int Magic;
-  int Version;
-  const char *Data;
-  const char *DataEnd;
-};
-
-template <typename T> T readAndAdvance(const char *&Ptr) {
-  auto V = *reinterpret_cast<const T *>(Ptr);
-  std::advance(Ptr, sizeof(T));
-  return V;
-}
-
-} // namespace
+#include "LanguageRegistration.h"
 
 static void readTUFatbin(const char *Binary, const FatbinWrapperTy *FW) {
   ol_device_handle_t Device = olKGetDefaultDevice();
@@ -143,10 +119,12 @@ void llvmRegisterManagedVar(void **, char *, char *, const char *, size_t,
                             unsigned) {
   fprintf(stderr, "RegisterManagedVar is not implemented!");
 }
+
 void llvmRegisterSurface(void **, const struct surfaceReference *,
                          const void **, const char *, int, int) {
   fprintf(stderr, "RegisterSurface is not implemented!");
 }
+
 void llvmRegisterTexture(void **, const struct textureReference *,
                          const void **, const char *, int, int, int) {
   fprintf(stderr, "RegisterTexture is not implemented!");
